apply icsp rules when mirroring images
- use icsp rules from ocs-registry image for mirroring images for
  disconnected deployment
- fixes: https://github.com/red-hat-storage/ocs-ci/issues/8103
